+++ src/polygon.h
@@ -4,7 +4,7 @@
  * https://github.com/fcambus/bdf2sfd
  *
  * Created:      2019-11-21
- * Last Updated: 2020-02-06
+ * Last Updated: 2020-02-07
  *
  * bdf2sfd is released under the BSD 2-Clause license
  * See LICENSE file for details
@@ -13,6 +13,6 @@
 #ifndef POLYGON_H
 #define POLYGON_H
 
-void polygon(uint32_t, uint32_t, uint32_t, int32_t, int32_t, uint32_t, uint32_t);
+void polygon(uint32_t, uint32_t, uint32_t, float, float, float, float);
 
 #endif /* POLYGON_H */
